Fix broken $PATH used to point to pipenv
You cannot use an 'export' in a RUN to create environment variable that
are referenced outside of the same shell running the RUN command. This
would mean the prefix which was meant to point to Pipenv would in fact
be empty, resulting in two /bin entries in $PATH.

Instead we just hard code it to the correct, known location.